fix map shape

+++ parse_map/read_map.c
@@ -6,7 +6,7 @@
 /*   By: ncontin <ncontin@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/28 11:16:33 by ncontin           #+#    #+#             */
-/*   Updated: 2025/02/03 12:44:11 by ncontin          ###   ########.fr       */
+/*   Updated: 2025/02/03 16:58:16 by ncontin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -47,10 +47,9 @@ void	read_map(char *arg, t_data *data, int fd)
 	if (fd < 0)
 		handle_open_error(data, map_path);
 	line = get_next_line(fd);
-	data->map->width = (int)ft_strlen(line) - 1;
+	data->map->width = ft_strlen(line) - 1;
 	while (line)
 	{
-		// check_map_shape(data, line, map_path);
 		free(line);
 		line = get_next_line(fd);
 		data->map->height++;
